LayoutTests/storage/indexeddb/modern/idbobjectstore-count-failures-expected.txt

This tests some obvious failures that can happen while calling IDBObjectStore.count().

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


ALERT: Initial upgrade needed: Old version - 0 New version - 1
ALERT: Failed to count records in object store with an invalid key
ALERT: Failed to count records in object store that's been deleted
ALERT: Initial upgrade versionchange transaction complete
ALERT: Failed to count records in object store while transaction is inactive
ALERT: readonly transaction complete
PASS successfullyParsed is true

TEST COMPLETE
